fix: Update container to humble-core
Somehow our general dependencies were not added to the container.

Also it tried to create the container based on the `iron-desktop-full`
variant, which is just big and doesn't help really, with all the usual
problems that come with it.

Therefore updated the container to the core image (which necessitates an
explicit `sudo` install).

Also we generally want the LTS version of ROS which currently is `humble`.